+++ examples/exmp017_roci/job.py
@@ -58,4 +58,15 @@ def run_exmp017(
 
     # > Parse JSON files
     output.parse()
+
+    # > Print S2 values
+    if s2_result := output.get_s2():
+        s2, ideal_s2 = s2_result
+        print(f"expectation S2: {s2}, ideal S**2: {ideal_s2}")
+    else:
+        print(f"Could not retrieve multiplicitiy from ORCA calculation: {output.get_outfile()}")
     return output
+
+
+if __name__ == "__main__":
+    run_exmp017()